No Email Selected Bug Fix
There was an issue where if the user selected the base "choose an email" option in the select menu, the page numbers wouldn't reset. Also if you tried to hit the page buttons, an error would come in the console as there's no object for it to read. I altered the event listener to check if the base select option was chosen, then if so clear the image and reset the page numbers. I also altered the page buttons to check if the base select option is selected, if so it doesn't allow the user to use the page buttons.