Email Regex + Email Added refinement
I changed the Email Regex to now also check that the user has added something like ".com" to the end of their email. I also clear input field now and scroll to the top of the page when a valid email is entered. Helps show the user they entered a valid email.