Javascript + S breakpoint adjustments
Changed the layout of the random image section at the smallest breakpoint. Now the email functionality is at the top of the screen to make for a better user experience. I also adjusted the sizing of the elements to work better on smaller screens. Lastly I modified my JS to work with the new HTML elements I implemented for the new layout.